asset/ : review architecture and cmake reworked
