+++ asset/include/asset/asset-cam.h
