+++ asset/include/asset/asset-computed.h
