+++ asset/include/asset/asset-configure-inform.h
