+++ asset/include/asset/asset-licensing.h
