+++ asset/include/asset/asset-manager.h
