+++ asset/include/asset/db/misc.h
