+++ asset/test/db2/misc.cpp
@@ -1,13 +1,35 @@
+/*  ====================================================================================================================
+    Copyright (C) 2014 - 2020 Eaton
+
+    This program is free software; you can redistribute it and/or modify
+    it under the terms of the GNU General Public License as published by
+    the Free Software Foundation; either version 2 of the License, or
+    (at your option) any later version.
+
+    This program is distributed in the hope that it will be useful,
+    but WITHOUT ANY WARRANTY; without even the implied warranty of
+    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+    GNU General Public License for more details.
+
+    You should have received a copy of the GNU General Public License along
+    with this program; if not, write to the Free Software Foundation, Inc.,
+    51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
+    ====================================================================================================================
+*/
+
 #include <asset/asset-db2.h>
-#include <catch2/catch.hpp>
 #include <fty_common_db_connection.h>
 #include <test-db/sample-db.h>
 #include <iostream>
 
-#define EXP_CHECK(expected)                                                                                                                \
-    if (!expected) {                                                                                                                       \
-        FAIL(expected.error());                                                                                                            \
-    }
+#include <catch2/catch.hpp>
+
+#define EXP_CHECK(expected)         \
+    do {                            \
+        if (!expected) {            \
+            FAIL(expected.error()); \
+        }                           \
+    } while(0)
 
 TEST_CASE("DB: misc")
 {
