+++ asset/test/export.cpp
@@ -1,8 +1,33 @@
+/*  ====================================================================================================================
+    Copyright (C) 2014 - 2020 Eaton
+
+    This program is free software; you can redistribute it and/or modify
+    it under the terms of the GNU General Public License as published by
+    the Free Software Foundation; either version 2 of the License, or
+    (at your option) any later version.
+
+    This program is distributed in the hope that it will be useful,
+    but WITHOUT ANY WARRANTY; without even the implied warranty of
+    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+    GNU General Public License for more details.
+
+    You should have received a copy of the GNU General Public License along
+    with this program; if not, write to the Free Software Foundation, Inc.,
+    51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
+    ====================================================================================================================
+*/
+
 #include "asset/asset-manager.h"
 #include <fty_common_db_connection.h>
-#include <catch2/catch.hpp>
 #include <fty/string-utils.h>
 #include <test-db/sample-db.h>
+
+#include <catch2/catch.hpp>
+
+#define REQUIRE_EXP(expr) \
+    do { \
+        if (!expr) { FAIL(expr.error().message()); } \
+    } while(0)
 
 using namespace fmt::literals;
